Moved the representation to extensions/ and add a plugin behaviour that registers the representation.
