+++ src/extensions/spherical-harmonic-surface/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Copyright (c) 2026 mol* contributors, licensed under MIT, See LICENSE file for more info.
+ *
+ * @author Ludovic Autin <autin@scripps.edu>
+ */
+
+export { SphericalHarmonicSurface } from './behavior';
