+++ src/mol-repr/structure/registry.ts
@@ -24,7 +24,6 @@ import { GaussianVolumeRepresentationProvider } from './representation/gaussian-
 import { BackboneRepresentationProvider } from './representation/backbone';
 import { PolyhedronRepresentationProvider } from './representation/polyhedron';
 import { PlaneRepresentationProvider } from './representation/plane';
-import { SphericalHarmonicSurfaceRepresentationProvider } from './representation/spherical-harmonic-surface';
 
 export class StructureRepresentationRegistry extends RepresentationRegistry<Structure, StructureRepresentationState> {
     constructor() {
@@ -54,7 +53,6 @@ export namespace StructureRepresentationRegistry {
         'putty': PuttyRepresentationProvider,
         'spacefill': SpacefillRepresentationProvider,
         'polyhedron': PolyhedronRepresentationProvider,
-        'spherical-harmonic-surface': SphericalHarmonicSurfaceRepresentationProvider,
     };
 
     type _BuiltIn = typeof BuiltIn
